+++ src/region.h
@@ -10,7 +10,7 @@ private:
 	int WIDTH;
 	int HEIGHT;
 	int FOV;
-	float shadowSmoothness;
+	double shadowSmoothness;
 	math::vec3 cameraPosition;
 	math::vec3 lightDirection;
 	std::vector<math::vec3> RMx;
@@ -19,17 +19,17 @@ private:
 
 	// get ray direction based on screen coordinates and apply
 	// yaw and pitch rotation to it
-	math::vec3 calculateRayDirection(float y, float x);
+	math::vec3 calculateRayDirection(double y, double x);
 
 	// marching!!!!
 	math::vec3 march(math::vec3 direction, int& step);
 	// in case the marched ray doesn't hit the fractal we render
 	// the background based on the pixel coordinates
-	math::vec3 renderBg(float y, float x);
+	math::vec3 renderBg(double y, double x);
 public:
 
 	region(int width, int height, fractal* f);
 	~region();
 
-	math::vec3 render(float y, float x);
+	math::vec3 render(double y, double x);
 };
